chore(mvn): move `it` profile to sw360 module
the tests can be executed via:
```
cd modules/sw360
mvn clean verify -DskipTests -P integration-test
```
