--- antenna-testing/antenna-sw360-integration-testing/src/test/resources/postgres/.gitignore
@@ -1,1 +1,0 @@
-*
