+++ modules/sw360/src/test/java/org/eclipse/sw360/antenna/sw360/integrationtesting/IntegrationTest.java
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) Bosch Software Innovations GmbH 2019.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.eclipse.sw360.antenna.sw360.integrationtesting;
+
+public class IntegrationTest {
+}
