+++ modules/sw360/src/test/resources/couchdb/data/.sw360db_design/mrview/129ec6bbb91373e3e947f79f55c34a07.view
