+++ modules/sw360/src/test/resources/couchdb/data/.sw360db_design/mrview/1c9257bd1f3a0f431dbeb02f05c14a9c.view
