+++ modules/sw360/src/test/resources/couchdb/data/.sw360db_design/mrview/bc879294dadcf39a2869674406496c9a.view
