+++ modules/sw360/src/test/resources/couchdb/data/.sw360users_design/mrview/3aea2376b78a4aaf3e421ee9d2cf23b6.view
