+++ modules/sw360/src/test/resources/couchdb/data/_users.couch
