+++ modules/sw360/src/test/resources/couchdb/data/sw360attachments.couch
