+++ modules/sw360/src/test/resources/docker-compose.yml
