+++ modules/sw360/src/test/resources/pom.xml
