+++ pom.xml
@@ -60,7 +60,6 @@
         <!-- assembly -->
         <module>assembly</module>
         <!-- misc -->
-        <module>antenna-testing</module>
         <module>antenna-license-provider</module>
         <module>antenna-csv-license-knowledge-base</module>
         <module>antenna-documentation</module>
